Checked for latest wp version 6.8, new block widget introduced, settings page updated fro more modern look

+++ uninstall.php
@@ -1,13 +1,18 @@
 <?php
-
 /**
- * Bail if uninstall constant is not defined
+ * Fired when the plugin is uninstalled.
+ *
+ * @since      2.0.0
+ * @package    Social_Sharer_For_Woo
+ * @author     Sajjad Hossain Sagor <sagorh672@gmail.com>
  */
 
-defined( 'WP_UNINSTALL_PLUGIN' ) || exit;
+// If uninstall not called from WordPress, then exit.
+if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+	die;
+}
 
 /**
  * Remove plugin options on uninstall/delete
  */
-
 delete_option( 'ssfwc_settings' );
